% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/anom.prec.R
\name{anom.prec}
\alias{anom.prec}
\title{Precipitation Anomoly Calculation}
\usage{
anom.prec(nc1, nc2)
}
\arguments{
\item{nc1}{a list object as created by \code{\link{retrieve.nc}},
\code{\link{extract.monthly.average}} or \code{\link{append.nc}}. The
minimum set of elements in the list must be dat, lat, lon and tim. See value
for a description of these elements.}

\item{nc2}{a list object as described in \code{nc1}.}
}
\value{
returns an object that is a list with the same elements as nc1 with
the \code{dat} element being the anomoly as a proportion of current.\cr\cr
the list will contain at least: \item{dat}{an array containing the anomoly
data. The dimensions of the array represent:\cr 1 - time described in this
list by \code{tim}\cr 2 - latitude (or y element) described in this list by
\code{lat}\cr 3 - longitude (or x element) described in this list by
\code{lon}} \item{lat}{vector of latitudes} \item{lon}{vector of longitudes}
\item{tim}{vector describing the time element either in e.g., julian days,
months or years}
}
\description{
\code{anom.prec} provides a method to estimate the anomoly as a
proportionate estimate of the future or past climate relative to the
current.
}
\examples{
##need to fill in
}
\author{
Jeremy VanDerWal \email{jjvanderwal@gmail.com}
}
\seealso{
\code{\link{anom.temp}}, \code{\link{retrieve.nc}},
\code{\link{extract.monthly.average}}, \code{\link{append.nc}}
}

