% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/averaging.R
\name{averaging}
\alias{averaging}
\title{Average the max temperature across a state}
\usage{
averaging(df, type = "mean")
}
\arguments{
\item{df}{Data frame containing a states max temperatures}

\item{type}{Do we take the mean or median?}
}
\value{
A data frame of averaged max temperature for each day
}
\description{
Average the max temperature across a state
}
