% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dap_bioclim.R
\name{dap_bioclim}
\alias{dap_bioclim}
\title{A wrapper function to execute bioclim in the climates package.}
\usage{
dap_bioclim(start, end, bbox_in, bioclims, OPeNDAP_URI, tmax_var, tmin_var,
  tave_var, prcp_var)
}
\arguments{
\item{start}{a start year}

\item{end}{an end year}

\item{bbox_in}{a bounding box to be passed to CF_bbox_in}

\item{bioclims}{a list of bioclims to be passed to the bioclim function.}

\item{OPeNDAP_URI}{an OPeNDAP resource without any service parameters.}

\item{tmax_var}{a string of the tmax variable id}

\item{tmin_var}{a string of the tmin variable id}

\item{tave_var}{optional: a string of the tave variable id}

\item{prcp_var}{a string of the prcp variable id}
}
\value{
A list of filenames that have been written to disk.
}
\description{
This function takes an input spatiotemporal domain and an OPeNDAP datasource.
It returns a list of geotiff filenames that have been written to the current working directory.
}
\examples{
Soon!
}

