% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dap_daily_stats.R
\name{dap_daily_stats}
\alias{dap_daily_stats}
\title{A wrapper function to execute daily_indices2geotiff in the climates package.}
\usage{
dap_daily_stats(start, end, bbox_in, thresholds, OPeNDAP_URI, tmax_var,
  tmin_var, tave_var, prcp_var, NetCDF_output = FALSE, fill_nas = FALSE)
}
\arguments{
\item{start}{a start year}

\item{end}{an end year}

\item{bbox_in}{a bounding box to be passed to CF_bbox_in}

\item{thresholds}{a named list to be passed to daily_indices}

\item{OPeNDAP_URI}{an OPeNDAP resource without any service parameters.}

\item{tmax_var}{a string of the tmax variable id}

\item{tmin_var}{a string of the tmin variable id}

\item{tave_var}{optional: a string of the tave variable id}

\item{prcp_var}{a string of the prcp variable id}

\item{NetCDF_output:}{Set to TRUE if NetCDF files are desired as output, otherwise will be geotiff.}

\item{Cells}{that are NA in only some time steps will be filled with the average of their neighbors.}
}
\value{
A list of filenames that have been written to disk.
}
\description{
This function takes an input spatiotemporal domain and an OPeNDAP datasource.
It returns a list of geotiff filenames that have been written to the current working directory.
}
\examples{
Soon!
}

