% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_dap_data.R
\name{get_dap_data}
\alias{get_dap_data}
\title{A function to retrieve OPeNDAP data for the climates package.}
\usage{
get_dap_data(ncdf4_handle, x1, y1, x2, y2, t_ind1, t_ind2, tmax_var, tmin_var,
  prcp_var, tave_var = NULL, temp_unit_func = NULL, fill_nas = FALSE)
}
\arguments{
\item{ncdf4_handle}{ncdf4 object.}

\item{x1}{x index from request_bbox function}

\item{y1}{y index from request_bbox function}

\item{x2}{x index from request_bbox function}

\item{y2}{y index from request_bbox function}

\item{t_ind1}{t index from request_time_bounds function}

\item{t_ind2}{t index from request_time_bounds function}

\item{tmax_var}{a string of the tmax variable id}

\item{tmin_var}{a string of the tmin variable id}

\item{prcp_var}{a string of the prcp variable id}

\item{tave_var}{optional: a string of the tave variable id}

\item{temp_unit_func}{function returned by init_dap function}

\item{Cells}{that are NA in only some time steps will be filled with the average of their neighbors.}
}
\value{
A named list of data according to inputs.
}
\description{
This function takes an ncdf4 object, indices to request, variables to request,
and a temperature convversion function. It returns cooresponding data.
}
\examples{
\dontrun{
Soon!
}
}

