% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_state.R
\name{get_state}
\alias{get_state}
\title{Get data for a given state at a given timeframe}
\usage{
get_state(filename, state = "oregon", timemin = 0, timemax = 14244)
}
\arguments{
\item{filename}{path for where data is located}

\item{state}{String for the state}

\item{timemin}{The start time}

\item{timemax}{The end time}
}
\value{
A data frame of the state data points within the time
}
\description{
Get data for a given state at a given timeframe
}
