% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_time_dim.R
\name{get_time_dim}
\alias{get_time_dim}
\title{A function to find the time dimension and units of a NetCDF-CF source.}
\usage{
get_time_dim(ncdf4_handle)
}
\arguments{
\item{ncdf4_handle}{ncdf4 object.}
}
\value{
A named list with time_units and the time_dim
}
\description{
This function takes an open ncdf4 data source.
It returns the time units and time dimension.
}
\examples{
Soon!
}

