% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/init_dap.R
\name{init_dap}
\alias{init_dap}
\title{An function to initialize an OPeNDAP resource for the climates package.}
\usage{
init_dap(OPeNDAP_URI, tmax_var, tmin_var, prcp_var, tave_var)
}
\arguments{
\item{OPeNDAP_URI}{an OPeNDAP resource without any service parameters.}

\item{tmax_var}{a string of the tmax variable id}

\item{tmin_var}{a string of the tmin variable id}

\item{prcp_var}{a string of the prcp variable id}
}
\value{
ncdf4_init a named list of an ncdf4 object and a temperature conversion function.
}
\description{
This function takes an OPeNDAP web service uri and a list of variables of interest.
It returns a named list containing the ncdf4 object's handle, and a temperature conversion function.
}
\examples{
Soon!
}

