% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/initialize_NetCDF.R
\name{initialize_NetCDF}
\alias{initialize_NetCDF}
\title{A function to initialize a NetCDF file for daily indices.}
\usage{
initialize_NetCDF(ncdf4_handle, thresholds, start = FALSE, end = FALSE,
  tmax_var = FALSE, prcp_var = FALSE, x_vals, y_vals, periods = list(),
  t_units = FALSE, p_units = FALSE)
}
\arguments{
\item{ncdf4_handle}{an initialized ncdf4 object.}

\item{thresholds}{a named list to be passed to daily_indices}

\item{start}{a start year}

\item{end}{an end year}

\item{tmax_var}{a string of the tmax variable id}

\item{prcp_var}{a string of the prcp variable id}

\item{periods}{a vector of years to use as time steps. (optional)}
}
\value{
A named list of including fileNames, x_vals, and y_vals for the NetCDF files that have been initiated.
}
\description{
This function initializes a set of NetCDF files for daily_indices. If start and end are given, it returns annual time steps.
If period is given, it creates time steps that span the years given in period. It returns a set of file names and the x/y coordinates.
}
\examples{
\dontrun{
Soon!
}
}

