% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/request_time_bounds.R
\name{request_time_bounds}
\alias{request_time_bounds}
\title{An OPeNDAP wrapper for the Climates function CF_date_range.}
\usage{
request_time_bounds(ncdf4_handle, start, end)
}
\arguments{
\item{ncdf4_handle}{ncdf4 object.}

\item{start}{a start year.}

\item{end}{an end year.}
}
\value{
time_indices a named list per CF_date_range.
}
\description{
This function takes an open ncdf4 data source, a start date and an end date.
It returns a named list per CF_date_range.
}
\examples{
\dontrun{
Soon!
}
}

