% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_filename.R
\name{get_filename}
\alias{get_filename}
\title{Get .nc filepath for data}
\usage{
get_filename(data = "ERA", specific = "", type = "PREC", year = "")
}
\arguments{
\item{data}{String containing data name, can be "ERA","Liu",or"CESM"}

\item{specific}{which dataset to pull - just for CESM and Liu}

\item{type}{Can be MAXT,MINT,PREC}
}
\value{
Filepath
}
\description{
Get .nc filepath for data
}
